+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `icon` support for `Audio` and `Video` attachments
 
+### Improved
+
+* Improve Interactions moderation
+* Compatibility with Akismet
+
 ## [4.4.0] - 2024-12-09
 
 ### Added